requirements.txt
django>=2.0
psycopg2>=2.5.4
gunicorn
djangorestframework
django-filter
